requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
asyncpg
alembic
psycopg2-binary
httpx
faker
requests
pytest
pytest-asyncio
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
black
isort
ruff